requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
sentence-transformers==2.7.0
transformers==4.36.0
huggingface_hub==0.19.4
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
torch==2.1.0
python-dotenv==1.0.0